+++ ffi/npm/lib/LineSegmenter.mjs
@@ -100,6 +100,24 @@ export class LineSegmenter {
     }
 
     /**
+     * Construct a {@link LineSegmenter} with default options (no locale-based tailoring) and no support for scripts requiring complex context dependent line breaks
+     * (Burmese, Khmer, Lao, and Thai), using compiled data
+     *
+     * See the [Rust documentation for `new_for_non_complex_scripts`](https://docs.rs/icu/2.1.1/icu/segmenter/struct.LineSegmenter.html#method.new_for_non_complex_scripts) for more information.
+     */
+    static createForNonComplexScripts() {
+
+        const result = wasm.icu4x_LineSegmenter_create_for_non_complex_scripts_mv1();
+
+        try {
+            return new LineSegmenter(diplomatRuntime.internalConstructor, result, []);
+        }
+
+        finally {
+        }
+    }
+
+    /**
      * Construct a {@link LineSegmenter} with custom options using compiled data. It automatically loads the best
      * available payload data for Burmese, Khmer, Lao, and Thai.
      *
@@ -236,6 +254,57 @@ export class LineSegmenter {
 
 
         const result = wasm.icu4x_LineSegmenter_create_dictionary_with_options_v2_and_provider_mv1(diplomatReceive.buffer, provider.ffiValue, contentLocale.ffiValue ?? 0, LineBreakOptions._fromSuppliedValue(diplomatRuntime.internalConstructor, options)._intoFFI(functionCleanupArena, {}, false));
+
+        try {
+            if (!diplomatReceive.resultFlag) {
+                const cause = new DataError(diplomatRuntime.internalConstructor, diplomatRuntime.enumDiscriminant(wasm, diplomatReceive.buffer));
+                throw new globalThis.Error('DataError.' + cause.value, { cause });
+            }
+            return new LineSegmenter(diplomatRuntime.internalConstructor, diplomatRuntime.ptrRead(wasm, diplomatReceive.buffer), []);
+        }
+
+        finally {
+            functionCleanupArena.free();
+
+            diplomatReceive.free();
+        }
+    }
+
+    /**
+     * Construct a {@link LineSegmenter} with custom options and no support for scripts requiring complex context dependent line breaks
+     * (Burmese, Khmer, Lao, and Thai), using compiled data.
+     *
+     * See the [Rust documentation for `new_for_non_complex_scripts`](https://docs.rs/icu/2.1.1/icu/segmenter/struct.LineSegmenter.html#method.new_for_non_complex_scripts) for more information.
+     */
+    static forNonComplexScriptsWithOptions(contentLocale, options) {
+        let functionCleanupArena = new diplomatRuntime.CleanupArena();
+
+
+        const result = wasm.icu4x_LineSegmenter_create_for_non_complex_scripts_with_options_v2_mv1(contentLocale.ffiValue ?? 0, LineBreakOptions._fromSuppliedValue(diplomatRuntime.internalConstructor, options)._intoFFI(functionCleanupArena, {}, false));
+
+        try {
+            return new LineSegmenter(diplomatRuntime.internalConstructor, result, []);
+        }
+
+        finally {
+            functionCleanupArena.free();
+
+        }
+    }
+
+    /**
+     * Construct a {@link LineSegmenter} with custom options and no support for complex languages
+     * (Burmese, Khmer, Lao, and Thai), using a particular data source.
+     *
+     * See the [Rust documentation for `new_for_non_complex_scripts`](https://docs.rs/icu/2.1.1/icu/segmenter/struct.LineSegmenter.html#method.new_for_non_complex_scripts) for more information.
+     */
+    static forNonComplexScriptsWithOptionsAndProvider(provider, contentLocale, options) {
+        let functionCleanupArena = new diplomatRuntime.CleanupArena();
+
+        const diplomatReceive = new diplomatRuntime.DiplomatReceiveBuf(wasm, 5, 4, true);
+
+
+        const result = wasm.icu4x_LineSegmenter_create_for_non_complex_scripts_with_options_v2_and_provider_mv1(diplomatReceive.buffer, provider.ffiValue, contentLocale.ffiValue ?? 0, LineBreakOptions._fromSuppliedValue(diplomatRuntime.internalConstructor, options)._intoFFI(functionCleanupArena, {}, false));
 
         try {
             if (!diplomatReceive.resultFlag) {