+++ provider/source/src/calendar/eras.rs
@@ -7,7 +7,6 @@ use crate::cldr_serde::eras::EraData;
 use crate::datetime::DatagenCalendar;
 use crate::SourceDataProvider;
 use icu::calendar::provider::*;
-use icu::calendar::types::Month;
 use icu::calendar::{AnyCalendar, Date};
 use icu_provider::prelude::*;
 use std::collections::BTreeMap;
@@ -22,7 +21,10 @@ impl SourceDataProvider {
     #[expect(clippy::type_complexity)]
     pub(crate) fn all_eras(
         &self,
-    ) -> Result<&BTreeMap<DatagenCalendar, Vec<(usize, EraData)>>, DataError> {
+    ) -> Result<
+        &BTreeMap<DatagenCalendar, (Option<DatagenCalendar>, Vec<(usize, EraData)>)>,
+        DataError,
+    > {
         let cldr = self.cldr()?;
         cldr.calendar_eras
             .get_or_init(|| {
@@ -58,72 +60,65 @@ impl SourceDataProvider {
                 ]
                 .into_iter()
                 .map(|cal| {
-                    let mut vec = if cal == DatagenCalendar::JapaneseExtended
-                        || cal == DatagenCalendar::JapaneseModern
-                    {
-                        era_dates_map[DatagenCalendar::Gregorian.cldr_name()]
-                            .clone()
-                            .eras
-                            .into_iter()
-                            .filter_map(|(id, data)| {
-                                data.code.as_ref()?;
-                                Some((id.parse::<usize>().ok()?, data))
-                            })
-                            .chain(
-                                era_dates_map[cal.cldr_name()]
-                                    .clone()
-                                    .eras
-                                    .into_iter()
-                                    .filter_map(|(key, mut data)| {
-                                        let key = key.parse::<usize>().ok()?;
-                                        if data.code.as_ref().is_none() {
-                                            if cal == DatagenCalendar::JapaneseExtended {
-                                                data.code =
-                                                    Some(crate::calendar::eras::era_to_code(
-                                                        japanese_names
-                                                            .get(&(key - 2).to_string())?,
-                                                        data.start?.year,
-                                                    ));
-                                            } else {
-                                                None?;
-                                            }
-                                        }
-                                        Some((key, data))
-                                    }),
-                            )
-                            .collect::<Vec<_>>()
-                    } else {
-                        let calendar =
-                            AnyCalendar::try_new_unstable(self, cal.canonical_any_calendar_kind())
-                                .unwrap();
+                    let inherit = era_dates_map[cal.cldr_name()]
+                        .inherit_eras
+                        .as_ref()
+                        .map(|c| DatagenCalendar::from_cldr_name(&c.calendar));
 
-                        era_dates_map[cal.cldr_name()]
-                            .clone()
-                            .eras
-                            .into_iter()
-                            .filter_map(|(key, mut data)| {
-                                let code = data.code.as_deref()?;
-                                // Check what ICU4X returns for the date 1-1-1 era
-                                data.icu4x_era_index = Date::try_new_from_codes(
-                                    Some(code),
-                                    1,
-                                    Month::new(1).code(),
-                                    1,
-                                    icu::calendar::Ref(&calendar),
-                                )
-                                .inspect_err(|e| {
-                                    log::warn!("Era '{code}' unknown by icu::calendar ({e:?})");
-                                })
-                                .ok()?
-                                .year()
-                                .era()?
-                                .era_index;
-                                Some((key.parse::<usize>().ok()?, data))
+                    struct EmptyProvider;
+                    impl<M: DataMarker<DataStruct = JapaneseEras<'static>>> DataProvider<M> for EmptyProvider {
+                        fn load(&self, _req: DataRequest) -> Result<DataResponse<M>, DataError> {
+                            Ok(DataResponse {
+                                metadata: Default::default(),
+                                payload: DataPayload::from_owned(
+                                    JapaneseEras {
+                                        dates_to_eras: [(EraStartDate { year: 3000, month: 1, day: 1 }, tinystr::tinystr!(16, "dummy"))].into_iter().collect(),
+                                    },
+                                ),
                             })
-                            .collect::<Vec<_>>()
-                    };
+                        }
+                    }
+
+                    let any_cal = AnyCalendar::try_new_unstable(&EmptyProvider, cal.canonical_any_calendar_kind()).unwrap();
+
+                    let mut vec = era_dates_map[cal.cldr_name()]
+                        .eras
+                        .iter()
+                        .filter_map(|(key, data)| {
+                            let mut data = data.clone();
+                            match cal {
+                                DatagenCalendar::JapaneseExtended => {
+                                    if data.code.is_empty() {
+                                        data.code = crate::calendar::eras::era_to_code(
+                                            japanese_names.get(&key.to_string())?,
+                                            data.start?.year,
+                                        );
+                                    }
+                                }
+                                _ => {
+                                    if cal == DatagenCalendar::JapaneseModern && data.start.unwrap().year < 1868 {
+                                        return None;
+                                    }
+                                    let date = data.start.or(data.end).unwrap();
+                                    let era_year =
+                                        Date::try_new_gregorian(date.year, date.month, date.day)
+                                            .unwrap()
+                                            .to_calendar(icu::calendar::Ref(&any_cal))
+                                            .year()
+                                            .era()
+                                            .unwrap();
+                                    if era_year.era != data.code {
+                                        println!("mismatched era code {era_year:?} - {data:?}");
+                                    }
+                                    data.icu4x_era_index = era_year.era_index;
+                                }
+                            }
+
+                            Some((key.parse::<usize>().unwrap(), data))
+                        })
+                        .collect::<Vec<_>>();
                     vec.sort_by_key(|&(k, _)| k);
-                    (cal, vec)
+                    (cal, (inherit, vec))
                 })
                 .collect())
             })
@@ -140,8 +135,6 @@ fn process_era_dates_map(
         core::mem::take(&mut data.get_mut("japanese").unwrap().eras)
             .into_iter()
             .map(|(idx, mut era)| {
-                // https://unicode-org.atlassian.net/browse/CLDR-18388 for why we need to do + 2
-                let idx = (idx.parse::<usize>().unwrap() + 2).to_string();
                 if let Some(start) = era.start.as_mut() {
                     // All pre-Taisho start dates are known to be wrong, this at least makes them valid.
                     // See https://unicode-org.atlassian.net/browse/CLDR-11400
@@ -160,36 +153,24 @@ fn process_era_dates_map(
     data
 }
 
-impl SourceDataProvider {
-    fn load_japanese_eras(
-        &self,
-        cal: DatagenCalendar,
-    ) -> Result<DataResponse<CalendarJapaneseModernV1>, DataError> {
-        let mut dates_to_eras = BTreeMap::new();
-
-        for (_, data) in self.all_eras()?[&cal].iter().skip(2) {
-            let start_date = data.start.unwrap();
-            let code = data.code.as_deref().unwrap();
-            let code = code.parse().map_err(|_| {
-                DataError::custom("Era code does not fit int TinyStr16").with_display_context(&code)
-            })?;
-
-            dates_to_eras.insert(start_date, code);
-        }
-
-        Ok(DataResponse {
-            metadata: Default::default(),
-            payload: DataPayload::from_owned(JapaneseEras {
-                dates_to_eras: dates_to_eras.into_iter().collect(),
-            }),
-        })
-    }
-}
-
 impl DataProvider<CalendarJapaneseModernV1> for SourceDataProvider {
     fn load(&self, req: DataRequest) -> Result<DataResponse<CalendarJapaneseModernV1>, DataError> {
         self.check_req::<CalendarJapaneseModernV1>(req)?;
-        self.load_japanese_eras(DatagenCalendar::JapaneseModern)
+
+        let (inherit, ref eras) = self.all_eras()?[&DatagenCalendar::JapaneseModern];
+
+        let dates_to_eras = inherit
+            .iter()
+            .flat_map(|i| self.all_eras().unwrap()[i].1.iter())
+            .chain(eras)
+            .filter(|(_, data)| !matches!(data.code.as_str(), "bce" | "ce"))
+            .map(|(_, data)| (data.start.unwrap(), data.code.parse().unwrap()))
+            .collect();
+
+        Ok(DataResponse {
+            metadata: Default::default(),
+            payload: DataPayload::from_owned(JapaneseEras { dates_to_eras }),
+        })
     }
 }
 
@@ -199,8 +180,18 @@ impl DataProvider<CalendarJapaneseExtendedV1> for SourceDataProvider {
         req: DataRequest,
     ) -> Result<DataResponse<CalendarJapaneseExtendedV1>, DataError> {
         self.check_req::<CalendarJapaneseExtendedV1>(req)?;
-        let DataResponse { metadata, payload } =
-            self.load_japanese_eras(DatagenCalendar::JapaneseExtended)?;
+
+        let (inherit, ref eras) = self.all_eras()?[&DatagenCalendar::JapaneseExtended];
+
+        let dates_to_eras = inherit
+            .iter()
+            .flat_map(|i| self.all_eras().unwrap()[i].1.iter())
+            .chain(eras)
+            .filter(|(_, data)| !matches!(data.code.as_str(), "bce" | "ce"))
+            .map(|(_, data)| (data.start.unwrap(), data.code.parse().unwrap()))
+            .collect();
+
+        let eras = JapaneseEras { dates_to_eras };
 
         // Integrity check
         //
@@ -211,7 +202,7 @@ impl DataProvider<CalendarJapaneseExtendedV1> for SourceDataProvider {
             let snapshot: JapaneseEras = serde_json::from_str(JAPANEXT_FILE)
                 .expect("Failed to parse the precached golden. This is a bug.");
 
-            if snapshot != *payload.get() {
+            if snapshot != eras {
                 return Err(DataError::custom(
                     "Era data has changed! This can be for two reasons: Either the CLDR locale data for Japanese eras has \
                     changed in an incompatible way, or there is a new Japanese era. Run \
@@ -224,8 +215,8 @@ impl DataProvider<CalendarJapaneseExtendedV1> for SourceDataProvider {
         }
 
         Ok(DataResponse {
-            metadata,
-            payload: payload.cast(),
+            metadata: Default::default(),
+            payload: DataPayload::from_owned(eras),
         })
     }
 }
@@ -279,6 +270,7 @@ impl crate::IterableDataProviderCached<CalendarJapaneseExtendedV1> for SourceDat
 #[test]
 pub fn ethiopic_and_ethioaa_are_compatible() {
     use icu::calendar::cal::{Ethiopian, EthiopianEraStyle};
+    use icu::calendar::types::Month;
     assert_eq!(
         Date::try_new_from_codes(
             Some("aa"),
@@ -306,22 +298,19 @@ pub fn ethiopic_and_ethioaa_are_compatible() {
 #[test]
 pub fn japanese_and_japanext_are_compatible() {
     let provider = SourceDataProvider::new_testing();
-    let japanese = &provider.all_eras().unwrap()[&DatagenCalendar::JapaneseModern];
-    let japanext = &provider.all_eras().unwrap()[&DatagenCalendar::JapaneseExtended];
+    let japanese = DataProvider::<CalendarJapaneseModernV1>::load(&provider, Default::default())
+        .unwrap()
+        .payload;
+    let japanext = DataProvider::<CalendarJapaneseExtendedV1>::load(&provider, Default::default())
+        .unwrap()
+        .payload;
     assert_eq!(
         japanext
+            .get()
+            .dates_to_eras
             .iter()
-            .take(2)
-            .zip(japanese.iter().take(2))
-            .find(|(e, a)| e != a),
-        None,
-    );
-    assert_eq!(
-        japanext
-            .iter()
-            .skip(2)
             .rev()
-            .zip(japanese.iter().skip(2).rev())
+            .zip(japanese.get().dates_to_eras.iter().rev())
             .find(|(e, a)| e != a),
         None,
         "{japanext:?} - {japanese:?}"
@@ -381,10 +370,9 @@ fn test_calendar_eras() {
             // Check that code and aliases produce identical results
             for era in era
                 .aliases
-                .as_deref()
-                .into_iter()
-                .flat_map(|s| s.split(' '))
-                .chain(era.code.as_deref())
+                .split(' ')
+                .chain(Some(era.code.as_str()))
+                .filter(|s| !s.is_empty())
             {
                 assert_eq!(
                     Date::try_new_from_codes(
@@ -409,8 +397,8 @@ fn test_calendar_eras() {
             }
 
             // Check that the correct era code is returned
-            if let Some(code) = era.code.as_deref() {
-                assert_eq!(era_year.era, code);
+            if !era.code.is_empty() {
+                assert_eq!(era_year.era, era.code);
             }
 
             // Check that the start/end date uses year 1, and minimal/maximal month/day