+++ CHANGELOG.md
@@ -9,6 +9,8 @@
         - Add conversion between `icu::properties::props::Script` and `icu::locale::subtags:Script` (unicode-org#7270)
         - Add conversion between `icu::properties::props::BidiClass` and `unicode_bidi::BidiClass` (unicode-org#7272)
         - Constify `PropertyNamesLong`/`PropertNamesShort`/`PropertyParser` constructors (unicode-org#7294)
+    - `icu_segmenter`
+        - Add non-complex line and word break constructors (unicode-org#7268)
 - Utils
   - Retire the `icu_harfbuzz` crate. The `icu_properties` and `icu_normalizer` types now directly implement the `harfbuzz-traits`
 
@@ -132,7 +134,7 @@
         - Document deprecated APIs (unicode-org#6890)
         - FFI for most new APIs
         - `ListFormatter::format` now takes a `diplomat::span<const diplomat::string_view_for_slice>` instead of a `diplomat::span<std::string_view>` to handle soundness issues on some platforms (unicode-org#6974)
-    - Dart version `2.1.0-dev.0`
+    - Dart version `2.1.0-dev.0`, `2.1.0-dev.1`
     - NPM version `2.1.0`
 - Utils
     - General